devbot/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
python-dotenv==0.14.0
requests==2.24.0
slackclient==2.7.3
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.4.2
